chore: moved useGetIsLhdAuth hook
